backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
boto3==1.28.25
lasio==0.31
pandas==2.0.3
numpy==1.24.3
google-generativeai==0.3.1
werkzeug==2.3.7
gunicorn==21.2.0
